+++ R/Get_DB_MIUR.R
@@ -222,10 +222,9 @@ Get_DB_MIUR <- function(Year = 2023, verbose = TRUE, input_Registry = NULL,
     YearMinus1 <- as.numeric(substr(year.patternA(Year),1,4))
     if(is.null(input_AdmUnNames)) {
       cat("Mapping cadastral codes to municipality (LAU) codes:")
-      input_AdmUnNames <- Get_AdmUnNames(
-        Year = ifelse(any(pattern %in% c(year.patternB(2016), year.patternB(2018))), Year, YearMinus1),
-        date = ifelse(any(pattern %in% c(year.patternB(2016), year.patternB(2018))), "01_01", "09_01"),
-        autoAbort = autoAbort)
+      AdmUnYear <- ifelse(any(pattern %in% c(year.patternB(2016), year.patternB(2018))), YearMinus1+1, YearMinus1)
+      AdmUnDay <- ifelse(any(pattern %in% c(year.patternB(2016), year.patternB(2018))), "01-01", "09-01")
+      input_AdmUnNames <- Get_AdmUnNames(Date = paste0(AdmUnYear, "-", AdmUnDay), autoAbort = autoAbort)
     }
     CodMun.R <- input_AdmUnNames %>% dplyr::select(.data$Cadastral_code, .data$Municipality_code)